Added control flow to bfmSpatial.
First, if a value for sensor is given, names(x) must correspond to Landsat sceneID's, otherwise this will be ignored with a warning message.

Second, if all values in a pixel stack are NA, it doesnt make sense to run bfastmonitor (and previously, this would return an error, and therefore a value of 1 to the error layer). Now, all output pixel values are NA if the original values for a pixel stack were all NA's (ie. if a mask had been previously applied).