Added support for z-dimension
This is implemented in functions where a dates vector is required (either explicitly supplied or derived from the z-dimension or names(x)).

Also, doMC was removed from @import and replaced with foreach, and also replaced in DESCRIPTION to prevent the error in roxygenise().

Package was finally roxygenise()'d after all changes